# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AGI Judd's Idea Catalog API",
    "description": "Transform LLM conversations into actionable Linear projects",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/conversations/{conversation_id}/project-draft/chat/stream": {
      "post": {
        "summary": "Stream Chat With Project Draft",
        "description": "Stream chat messages with real-time updates via Server-Sent Events.",
        "operationId": "stream_chat_with_project_draft_api_conversations__conversation_id__project_draft_chat_stream_post",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/conversations": {
      "get": {
        "summary": "List Conversations",
        "description": "Get a paginated list of all imported conversations.",
        "operationId": "list_conversations_api_conversations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/ConversationListResponse"
                    },
                    {
                      "$ref": "#/components/schemas/ErrorResponse"
                    }
                  ],
                  "title": "Response List Conversations Api Conversations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "minLength": 1,
            "title": "Message",
            "description": "User message content"
          },
          "llm_provider": {
            "type": "string",
            "minLength": 1,
            "title": "Llm Provider",
            "description": "LLM provider to use"
          },
          "llm_model": {
            "type": "string",
            "minLength": 1,
            "title": "Llm Model",
            "description": "LLM model to use"
          },
          "attachment_ids": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "title": "Attachment Ids",
            "description": "List of file attachment IDs to include with this message"
          }
        },
        "type": "object",
        "required": [
          "message",
          "llm_provider",
          "llm_model"
        ],
        "title": "ChatRequest",
        "description": "Request model for sending a chat message."
      },
      "ConversationListItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "import_date": {
            "type": "string",
            "title": "Import Date"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "is_locked": {
            "type": "boolean",
            "title": "Is Locked"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "user_name": {
            "type": "string",
            "title": "User Name"
          },
          "user_email": {
            "type": "string",
            "title": "User Email"
          },
          "project_draft_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Draft Title"
          },
          "project_draft_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Draft Description"
          },
          "linear_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Linear Url"
          },
          "last_user_message_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last User Message Content"
          },
          "last_assistant_message_content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Assistant Message Content"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "title",
          "import_date",
          "created_at",
          "updated_at",
          "is_locked",
          "user_id",
          "user_name",
          "user_email"
        ],
        "title": "ConversationListItem",
        "description": "Response model for dashboard list view."
      },
      "ConversationListResponse": {
        "properties": {
          "conversations": {
            "items": {
              "$ref": "#/components/schemas/ConversationListItem"
            },
            "type": "array",
            "title": "Conversations",
            "description": "List of conversations"
          }
        },
        "type": "object",
        "required": [
          "conversations"
        ],
        "title": "ConversationListResponse",
        "description": "Response for listing conversations."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Error message"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Additional error details"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "Standard error response schema."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}